sayHi4

+++ src/main/java/com/company/Calculator.java
@@ -12,7 +12,7 @@ public class Calculator {
     }
 
     public long multiply(int a, int b) {
-        return a * b;
+        return (long) a * b;
     }
 
     public double division(int a, int b) {
